requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
antlr4-python3-runtime==4.9.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
braceexpand==0.1.7
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
charset-normalizer==3.4.7
click==8.1.8
cloudpickle==3.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1729602398863/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
decorator==5.2.1
faster-fifo==1.5.2
filelock==3.19.1
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1755224029921/work
fsspec==2025.10.0
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1753444155689/work
grpcio==1.80.0
gym==0.23.0
gym-notices==0.1.0
-e git+https://github.com/facebookresearch/habitat-lab.git@094d6be2f9d057e4781a68ae792132895fd4d3d0#egg=habitat_baselines&subdirectory=habitat-baselines
-e git+https://github.com/facebookresearch/habitat-lab.git@094d6be2f9d057e4781a68ae792132895fd4d3d0#egg=habitat_lab&subdirectory=habitat-lab
habitat_sim==0.3.3
hf-xet==1.4.3
huggingface_hub==0.36.2
hydra-core==1.3.2
idna==3.13
ifcfg==0.24
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1738273805233/work
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
Jinja2==3.1.6
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1725459226886/work
llvmlite==0.43.0
lmdb==2.2.0
magnum @ file:///Users/runner/miniconda3/envs/habitat/conda-bld/habitat-sim_1738007552115/work/build/deps/magnum-bindings/src/python
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
munkres==1.1.4
networkx==3.2.1
numba @ file:///Users/runner/miniforge3/conda-bld/numba_1718888154750/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225689808/work/dist/numpy-1.26.4-cp39-cp39-macosx_11_0_arm64.whl#sha256=989042d872faf31c0434bd99fec5ec09b5c5d9a0b8125edd0ae869c5d8851048
numpy-quaternion @ file:///Users/runner/miniforge3/conda-bld/quaternion_1719301757253/work
objectio==0.2.29
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1777103621/work
pillow==10.4.0
proglog==0.1.12
protobuf==6.33.6
Pygments==2.20.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1753873557/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==15.0.0
safetensors==0.7.0
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1716470312029/work/dist/scipy-1.13.1-cp39-cp39-macosx_11_0_arm64.whl#sha256=95cc9a6aceed50037865da4a45baac05fd0dcee7e557c6149fde270f2ffebafc
shellingham==1.5.4
simplejson==4.1.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1611376390914/work
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
torchvision==0.23.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1754732067879/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers==4.57.6
typer==0.23.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1736692557404/work
urllib3==2.6.3
webdataset==0.1.40
Werkzeug==3.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
